bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.2.1
flask
gigachat